+++ tests/testthat/test-category.R
@@ -44,3 +44,20 @@ test_that("season splits work under all circumstances", {
   expect_equal(cat_res$season[71], "Year-round")
   expect_equal(cat_res$season[72], "Fall-Spring")
 })
+
+test_that("y = any existing column", {
+  ts <- ts2clm(sst_Med, climatologyPeriod = c("1983-01-01", "2012-12-31"))
+  res <- detect_event(data = ts)
+  res$climatology$pawpaw <- res$climatology$temp
+  expect_is(category(res, y = pawpaw), "tbl_df")
+})
+
+test_that("climatology = T causes a list output with the time series category data", {
+  res <- detect_event(ts2clm(sst_Med, climatologyPeriod = c("1983-01-01", "2012-12-31")))
+  cat <- category(res, climatology = T)
+  expect_is(cat, "list")
+  expect_is(cat$climatology, "tbl_df")
+  expect_is(cat$event, "tbl_df")
+  expect_equal(ncol(cat$climatology), 12)
+  expect_equal(nrow(cat$climatology), 900)
+})