+++ tests/testthat/test-ts2clm.R
@@ -73,3 +73,14 @@ test_that("mssing data causes na_interp() to be used", {
   expect_lt(ts_2$temp[400], ts_1$temp[400])
   expect_lt(ts_2$temp[1000], ts_1$temp[1000])
 })
+
+test_that("contiguous mssing data causes clim_calc() to be used", {
+  sst_WA_cont <- sst_WA %>%
+    dplyr::mutate(month = month(t)) %>%
+    dplyr::filter(month != 1) %>%
+    dplyr::select(-month)
+  res <- ts2clm(sst_WA_cont, climatologyPeriod = c("1983-01-01", "2012-12-31"))
+  expect_is(res, "data.frame")
+  expect_equal(ncol(res), 6)
+  expect_equal(nrow(res), 12022)
+})